src/part2/services/text_indexer/requirements.txt
kafka-python>=2.0
sentence-transformers>=3.0
faiss-cpu>=1.8.0
numpy>=1.26
